nr12_erp/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.0
python-dotenv>=1.0
Pillow>=10.0
qrcode>=7.4
python-telegram-bot>=21.0,<22.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
dj-database-url>=2.1.0
